gestion de session

+++ assets/javascript/login.js
@@ -10,20 +10,22 @@ document.getElementById("submit-login").addEventListener("click", connexion);
 function connexion(event){
     const email = document.getElementById("email");
     const password = document.getElementById("password");
-
     const dataSession = {
         email: email.value,
         password: password.value,
-    }
-
-    sessionStorage.setItem("ADMIN", JSON.stringify(dataSession));
-    const dataLocal = JSON.parse(localStorage.getItem("ADMINISTRATEURS")).find(cle => cle.email ==dataSession.email && cle.password == dataSession.password);
-    if (dataLocal){
-        window.location.href = './../corporates/dashboard.html';
+    }    
+    if (localStorage.getItem("ADMINISTRATEURS")){
+        sessionStorage.setItem("ADMIN", JSON.stringify(dataSession));
+        const dataLocal = JSON.parse(localStorage.getItem("ADMINISTRATEURS")).find(cle => cle.email == dataSession.email && cle.password == dataSession.password);
+        if (dataLocal) {
+            window.location.href = "https://nfcdjobo.github.io/Pro-Gest-All/corporates/dashboard.html";
+        }else{
+            sessionStorage.clear();
+            document.getElementById("error-password").textContent = "Email ou mot de passe incorrecte."
+        }
     }else{
-        sessionStorage.clear();
-        document.getElementById("error-password").textContent = "Email ou mot de passe incorrecte."
-
+        alert("Ce utilisateur n'est pas reconnu. Veuillez donc créer un compte.");
+        window.location.href = "https://nfcdjobo.github.io/Pro-Gest-All/corporates/sinscrire.html";
     }
     
 }
